build(fix): typos and swaps (#3)
* build: use swap file

Needed this due to a small amount of memory that is available on
t3.micro instance

* build: get all systems libraries and install R packages properly

Needed certain system libraries like fribidi for all packages to install
properly, also BiocManager is required as ggtranscript doesn't download
dependencies properly as it is written poorly.

* build(fix): incorrect package name (typo)

* build: Split commands and increase timeout

Due to packages being very large, the default of 10 minutes was too
short a time, leading to timeouts. Splitting up the R installation
should help with this. Giving each command a full hour to finish.

* build(fix): Only create swapfile once

Commands will fail if swapfile already exists